+++ main-service/src/main/java/ru/practicum/category/repository/CategoryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import ru.practicum.category.model.Category;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Long>, QuerydslPredicateExecutor<Category> {
+public interface CategoryRepository extends JpaRepository<Category, Integer>, QuerydslPredicateExecutor<Category> {
 
 }
