+++ main-service/src/main/java/ru/practicum/event/dto/UpdateEventUserRequest.java
@@ -14,7 +14,7 @@ public class UpdateEventUserRequest {
     @Size(min = 20, max = 2000)
     private String annotation;
 
-    private Long category;
+    private Integer category;
 
     @Size(min = 20, max = 7000)
     private String description;
