+++ main-service/src/main/java/ru/practicum/event/service/AdminEventService.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface AdminEventService {
 
     List<EventFullDto> getEventsForAdmin(Pageable pageable, Integer offset, Integer size, List<Long> usersIds,
-                                         List<String> states, List<Long> categoriesIds, LocalDateTime rangeEnd, LocalDateTime rangeStart);
+                                         List<String> states, List<Integer> categoriesIds, LocalDateTime rangeEnd, LocalDateTime rangeStart);
 
     EventFullDto updateEvent(Long eventId, UpdateEventAdminRequest updateEventAdminRequest);
 }
